+++ apps/web/app/page.tsx
@@ -2,81 +2,190 @@ import Link from "next/link";
 import {
   Card,
   CardContent,
+  CardHeader,
+  CardTitle,
+  CardDescription,
 } from "@repo/ui/components/ui/card";
-import { Navbar, HeroSection } from "./_components/navbar-hero";
-import { AuthCard } from "./_components/auth-card";
+import { Button } from "@repo/ui/components/ui/button";
+import { Separator } from "@repo/ui/components/ui/separator";
+import {
+  ArrowRight,
+  Sparkles,
+  GitBranch,
+  BarChart2,
+  ChevronRight
+} from "lucide-react";
 
-/* ── SVG Icons (static, no interactivity needed) ─────────────────── */
+/* ── Brand Icons (since lucide lacks them) ───────────────────────── */
 
-function SparklesIcon({ className }: { className?: string }) {
+function GoogleIcon({ className }: { className?: string }) {
   return (
-    <svg
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      className={className ?? "w-5 h-5"}
-    >
-      <path d="M9.937 15.5A2 2 0 0 0 8.5 14.063l-6.135-1.582a.5.5 0 0 1 0-.962L8.5 9.936A2 2 0 0 0 9.937 8.5l1.582-6.135a.5.5 0 0 1 .963 0L14.063 8.5A2 2 0 0 0 15.5 9.937l6.135 1.581a.5.5 0 0 1 0 .964L15.5 14.063a2 2 0 0 0-1.437 1.437l-1.582 6.135a.5.5 0 0 1-.963 0z" />
-      <path d="M20 3v4" />
-      <path d="M22 5h-4" />
-      <path d="M4 17v2" />
-      <path d="M5 18H3" />
+    <svg viewBox="0 0 24 24" fill="currentColor" className={className ?? "w-5 h-5"}>
+      <path d="M12.48 10.92v3.28h7.84c-.24 1.84-.853 3.187-1.787 4.133-1.147 1.147-2.933 2.4-6.053 2.4-4.827 0-8.6-3.893-8.6-8.72s3.773-8.72 8.6-8.72c2.6 0 4.507 1.027 5.907 2.347l2.307-2.307C18.747 1.44 16.133 0 12.48 0 5.867 0 .307 5.387.307 12s5.56 12 12.173 12c3.573 0 6.267-1.173 8.373-3.36 2.16-2.16 2.84-5.213 2.84-7.667 0-.76-.053-1.467-.173-2.053H12.48z" />
     </svg>
   );
 }
 
-function BranchIcon({ className }: { className?: string }) {
+function GithubIcon({ className }: { className?: string }) {
   return (
-    <svg
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      className={className ?? "w-5 h-5"}
-    >
-      <path d="M6 3v12" />
-      <circle cx="18" cy="6" r="3" />
-      <circle cx="6" cy="18" r="3" />
-      <path d="M18 9a9 9 0 0 1-9 9" />
+    <svg viewBox="0 0 24 24" fill="currentColor" className={className ?? "w-5 h-5"}>
+      <path d="M12 0C5.37 0 0 5.37 0 12c0 5.31 3.435 9.795 8.205 11.385.6.105.825-.255.825-.57 0-.285-.015-1.23-.015-2.235-3.015.555-3.795-.735-4.035-1.41-.135-.345-.72-1.41-1.23-1.695-.42-.225-1.02-.78-.015-.795.945-.015 1.62.87 1.845 1.23 1.08 1.815 2.805 1.305 3.495.99.105-.78.42-1.305.765-1.605-2.67-.3-5.46-1.335-5.46-5.925 0-1.305.465-2.385 1.23-3.225-.12-.3-.54-1.53.12-3.18 0 0 1.005-.315 3.3 1.23.96-.27 1.98-.405 3-.405s2.04.135 3 .405c2.295-1.56 3.3-1.23 3.3-1.23.66 1.65.24 2.88.12 3.18.765.84 1.23 1.905 1.23 3.225 0 4.605-2.805 5.625-5.475 5.925.435.375.81 1.095.81 2.22 0 1.605-.015 2.895-.015 3.3 0 .315.225.69.825.57A12.02 12.02 0 0 0 24 12c0-6.63-5.37-12-12-12z" />
     </svg>
   );
 }
 
-function ChartIcon({ className }: { className?: string }) {
+/* ── Components Defined Directly in page.tsx ─────────────────────── */
+
+function Navbar() {
   return (
-    <svg
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      className={className ?? "w-5 h-5"}
-    >
-      <path d="M3 3v16a2 2 0 0 0 2 2h16" />
-      <path d="m19 9-5 5-4-4-3 3" />
-    </svg>
+    <header className="fixed top-0 left-0 w-full z-50 bg-background/80 backdrop-blur-md border-b border-border">
+      <div className="max-w-7xl mx-auto flex items-center justify-between h-16 px-6 lg:px-8">
+        <div className="flex items-center gap-8">
+          <Link
+            href="/"
+            className="text-xl font-bold font-[family-name:var(--font-space-grotesk)] tracking-tight"
+          >
+            Snap-Form
+          </Link>
+          <nav className="hidden md:flex items-center gap-1">
+            <a
+              href="#features"
+              className="px-3 py-2 text-sm text-muted-foreground hover:text-foreground transition-colors rounded-lg hover:bg-accent"
+            >
+              Features
+            </a>
+            <a
+              href="#how-it-works"
+              className="px-3 py-2 text-sm text-muted-foreground hover:text-foreground transition-colors rounded-lg hover:bg-accent"
+            >
+              How It Works
+            </a>
+            <a
+              href="#get-started"
+              className="px-3 py-2 text-sm text-muted-foreground hover:text-foreground transition-colors rounded-lg hover:bg-accent"
+            >
+              Get Started
+            </a>
+          </nav>
+        </div>
+        <div className="flex items-center gap-3">
+          <Button variant="ghost" size="sm" render={<a href="#get-started" />}>
+            Sign In
+          </Button>
+          <Button size="sm" render={<a href="#get-started" />}>
+            Get Started
+            <ArrowRight className="w-3.5 h-3.5" />
+          </Button>
+        </div>
+      </div>
+    </header>
   );
 }
 
-function ChevronRightIcon({ className }: { className?: string }) {
+function HeroSection() {
   return (
-    <svg
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      className={className ?? "w-4 h-4"}
-    >
-      <path d="m9 18 6-6-6-6" />
-    </svg>
+    <section className="w-full flex flex-col items-center text-center pt-20 md:pt-32 pb-16 md:pb-24 px-6 lg:px-8 max-w-5xl mx-auto">
+      <div className="inline-flex items-center gap-2 px-4 py-1.5 mb-8 rounded-full border border-border bg-accent/50 text-sm text-muted-foreground">
+        <Sparkles className="w-3.5 h-3.5 text-foreground" />
+        <span>AI-native form builder</span>
+      </div>
+
+      <h1 className="text-5xl md:text-7xl font-bold font-[family-name:var(--font-space-grotesk)] tracking-[-0.03em] leading-[1.1] max-w-4xl">
+        Forms at the speed
+        <br />
+        of thought.
+      </h1>
+
+      <p className="mt-6 text-lg md:text-xl text-muted-foreground max-w-2xl leading-relaxed">
+        The AI-native form builder for technical teams. Describe your logic,
+        and let the engine generate production-ready interfaces instantly.
+      </p>
+
+      {/* AI Prompt Pill */}
+      <div className="w-full max-w-2xl mt-10 relative group">
+        <div className="absolute inset-0 bg-primary/5 blur-xl group-hover:bg-primary/10 transition-colors duration-500 rounded-full" />
+        <a
+          href="#get-started"
+          className="relative flex items-center bg-background border border-border rounded-full p-1.5 shadow-sm transition-all hover:ring-2 hover:ring-ring hover:ring-offset-2 hover:ring-offset-background cursor-pointer"
+        >
+          <Sparkles className="ml-4 mr-2 text-muted-foreground w-5 h-5 shrink-0" />
+          <span className="flex-grow text-left text-muted-foreground text-base h-11 flex items-center select-none">
+            Describe your form...
+          </span>
+          <Button
+            className="rounded-full h-10 px-6 shrink-0 pointer-events-none"
+            tabIndex={-1}
+            aria-hidden
+          >
+            Generate
+            <ArrowRight className="w-3.5 h-3.5" />
+          </Button>
+        </a>
+      </div>
+
+      {/* Secondary Actions */}
+      <div className="flex flex-col sm:flex-row gap-3 mt-8">
+        <Button variant="outline" size="lg" render={<a href="#get-started" />}>
+          Get Started
+        </Button>
+      </div>
+    </section>
+  );
+}
+
+function AuthCard() {
+  return (
+    <Card className="w-full max-w-md">
+      <CardHeader className="text-center">
+        <CardTitle className="text-lg">Welcome to Snap-Form</CardTitle>
+        <CardDescription>
+          Choose your preferred method to continue
+        </CardDescription>
+      </CardHeader>
+      <CardContent className="space-y-4">
+        {/* Mock Buttons — no onClick handlers since this is a server component */}
+        <Button
+          id="google-login-btn"
+          variant="outline"
+          size="lg"
+          className="w-full justify-center gap-3"
+          render={<a href="#get-started" />}
+        >
+          <GoogleIcon className="w-5 h-5" />
+          Continue with Google
+        </Button>
+
+        <Button
+          id="github-login-btn"
+          variant="outline"
+          size="lg"
+          className="w-full justify-center gap-3"
+          render={<a href="#get-started" />}
+        >
+          <GithubIcon className="w-5 h-5" />
+          Continue with GitHub
+        </Button>
+
+        <div className="relative py-2">
+          <Separator />
+          <span className="absolute top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2 bg-card px-3 text-xs text-muted-foreground">
+            or
+          </span>
+        </div>
+
+        <p className="text-xs text-center text-muted-foreground leading-relaxed">
+          By continuing, you agree to our{" "}
+          <Link href="/terms" className="underline hover:text-foreground">
+            Terms of Service
+          </Link>{" "}
+          and{" "}
+          <Link href="/privacy" className="underline hover:text-foreground">
+            Privacy Policy
+          </Link>
+          .
+        </p>
+      </CardContent>
+    </Card>
   );
 }
 
@@ -85,12 +194,12 @@ function ChevronRightIcon({ className }: { className?: string }) {
 export default function Home() {
   return (
     <div className="bg-background text-foreground min-h-screen flex flex-col selection:bg-primary selection:text-primary-foreground">
-      {/* ─── Navbar (Client) ────────────────────────────────────── */}
+      {/* ─── Navbar ────────────────────────────────────────────── */}
       <Navbar />
 
       {/* ─── Main Content ───────────────────────────────────────── */}
       <main className="flex-grow pt-16">
-        {/* ── Hero Section (Client) ─────────────────────────────── */}
+        {/* ── Hero Section ─────────────────────────────── */}
         <HeroSection />
 
         {/* ── Social Proof ───────────────────────────────────────── */}
@@ -139,7 +248,7 @@ export default function Home() {
             <Card className="group hover:border-primary/50 transition-colors duration-300">
               <CardContent className="p-8 flex flex-col gap-6 h-full">
                 <div className="w-12 h-12 rounded-xl border border-border bg-accent flex items-center justify-center group-hover:bg-primary group-hover:text-primary-foreground group-hover:border-primary transition-all duration-300">
-                  <SparklesIcon className="w-5 h-5" />
+                  <Sparkles className="w-5 h-5" />
                 </div>
                 <div>
                   <h3 className="font-semibold font-[family-name:var(--font-space-grotesk)] text-lg mb-2">
@@ -153,7 +262,7 @@ export default function Home() {
                 <div className="mt-auto pt-6 border-t border-border">
                   <span className="inline-flex items-center text-sm font-medium text-foreground cursor-pointer hover:underline">
                     Explore syntax
-                    <ChevronRightIcon className="w-3.5 h-3.5 ml-1" />
+                    <ChevronRight className="w-3.5 h-3.5 ml-1" />
                   </span>
                 </div>
               </CardContent>
@@ -163,7 +272,7 @@ export default function Home() {
             <Card className="group hover:border-primary/50 transition-colors duration-300">
               <CardContent className="p-8 flex flex-col gap-6 h-full">
                 <div className="w-12 h-12 rounded-xl border border-border bg-accent flex items-center justify-center group-hover:bg-primary group-hover:text-primary-foreground group-hover:border-primary transition-all duration-300">
-                  <BranchIcon className="w-5 h-5" />
+                  <GitBranch className="w-5 h-5" />
                 </div>
                 <div>
                   <h3 className="font-semibold font-[family-name:var(--font-space-grotesk)] text-lg mb-2">
@@ -175,7 +284,7 @@ export default function Home() {
                   </p>
                 </div>
                 <div className="mt-auto pt-6 border-t border-border w-full h-24 bg-accent/50 rounded-lg flex items-center justify-center relative overflow-hidden">
-                  <BranchIcon className="w-16 h-16 text-border absolute" />
+                  <GitBranch className="w-16 h-16 text-border absolute" />
                 </div>
               </CardContent>
             </Card>
@@ -184,7 +293,7 @@ export default function Home() {
             <Card className="group hover:border-primary/50 transition-colors duration-300">
               <CardContent className="p-8 flex flex-col gap-6 h-full">
                 <div className="w-12 h-12 rounded-xl border border-border bg-accent flex items-center justify-center group-hover:bg-primary group-hover:text-primary-foreground group-hover:border-primary transition-all duration-300">
-                  <ChartIcon className="w-5 h-5" />
+                  <BarChart2 className="w-5 h-5" />
                 </div>
                 <div>
                   <h3 className="font-semibold font-[family-name:var(--font-space-grotesk)] text-lg mb-2">
@@ -281,7 +390,7 @@ export default function Home() {
             </p>
           </div>
 
-          {/* Auth Card (Client) */}
+          {/* Auth Card */}
           <AuthCard />
         </section>
       </main>
