+++ apps/web/package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@repo/types": "workspace:*",
     "@repo/ui": "*",
+    "lucide-react": "^1.23.0",
     "next": "16.2.9",
     "react": "19.2.4",
     "react-dom": "19.2.4"
